Добавляет роутинг
