+++ src/components/404.tsx
@@ -1,0 +1,6 @@
+export const NotFound = () => (
+  <div style={{margin: '0 auto', textAlign: 'center'}}>
+    <p>Упс, страница не найдена</p>
+    <a href='/'>Вернуться на главную</a>
+  </div>
+);
