--- src/components/App.tsx
@@ -1,5 +1,0 @@
-import {Main} from './Main.tsx';
-
-export const App = ({props}: {props: number}) => (
-  <Main offersCount={props}/>
-);
