+++ src/pages/favorites-empty.tsx
@@ -1,5 +1,5 @@
 
-export const FavoritesEmpty = () => (
+export const FavoritesEmpty = ():JSX.Element => (
   <main className="page__main page__main--favorites page__main--favorites-empty">
     <div className="page__favorites-container container">
       <section className="favorites favorites--empty">
